ekump/APMSP-1076 break up trace utils send data file (#471)
* Move SendDataResult and RetryStrategy for trace_utils::SendDataResult

to separate files and add rustdoc commments for public functions.

* minor refactor of trace_utils::SendData to reduce send_payload length

No logic changes to refactor, just moving some logic to separate
functions for readability. Also, reduce visibility of SendDataResult.

* Remove no longer relevant TODO for retry logic in sidecar trace flusher

* Reduce visibility of trace_utils::send_data::RetryStrategy struct fields

Added a new() impl and max_retries() function.
